op-deployer: refactor main to make more testable
